requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
rembg==2.0.62
gunicorn
Pillow
numpy
onnxruntime